+++ jobs/CHANGELOG.rst
@@ -6,6 +6,7 @@ This document describes changes between each past release.
 1.1.2 (unreleased)
 ------------------
 
+- Fix event handling of RC metadata (fixes #314)
 - Fix exclusion of thunderbird nightly releases (fixes #312)
 
 1.1.1 (2017-11-30)